replaceData {DT}	R Documentation

Replace data in an existing table

Description

Replace the data object of a table output and avoid regenerating the full table, in which case the state of the current table will be preserved (sorting, filtering, and pagination) and applied to the table with new data.

Usage


replaceData(proxy, data, ..., resetPaging = TRUE, clearSelection = "all")

Arguments

proxy
a proxy object created by dataTableProxy()

data
the new data object to be loaded in the table

...
other arguments to be passed to dataTableAjax()

resetPaging, clearSelection
passed to reloadData()

Note

When you replace the data in an existing table, please make sure the new data has the same number of columns as the current data. When you have enabled column filters, you should also make sure the attributes of every column remain the same, e.g. factor columns should have the same or fewer levels, and numeric columns should have the same or smaller range, otherwise the filters may never be able to reach certain rows in the data.

[Package DT version 0.9 Index]